requirement.txt
Django==1.11.8
django-js-asset==1.1.0
django-mptt==0.9.1
djangorestframework==3.6.2
MySQL-python==1.2.5
pytz==2018.5
